[#104] refactor: category 조회 시 join 이 아닌 쿼리로 변경

+++ live-commerce-api/src/main/java/org/giglab/live/commerce/api/dto/campaign/CampaignSummaryItem.java
@@ -8,4 +8,4 @@ public record CampaignSummaryItem(
     String description,
     String status,
     LocalDateTime scheduledAt,
-    int productCount) {}
+    long productCount) {}
