+++ live-commerce-core/src/main/java/org/giglab/live/commerce/core/campaign/application/dto/CampaignSummary.java
@@ -9,4 +9,4 @@ public record CampaignSummary(
     String description,
     BroadcastStatusType status,
     LocalDateTime scheduledAt,
-    int productCount) {}
+    long productCount) {}
